renamed all constant files configured project under GeoDjango + postgis
