+++ project/authentication/services.py
@@ -3,7 +3,7 @@ import string
 from datetime import date
 from typing import Any
 
-from authentication.constant.code_types import (
+from authentication.constants.code_types import (
     ACCOUNT_DELETE_CODE_TYPE,
     EMAIL_CHANGE_CODE_TYPE,
     EMAIL_VERIFY_CODE_TYPE,
@@ -11,7 +11,7 @@ from authentication.constant.code_types import (
     PASSWORD_RESET_CODE_TYPE,
     PHONE_CHANGE_CODE_TYPE,
 )
-from authentication.constant.success import (
+from authentication.constants.success import (
     EMAIL_MESSAGE_TEMPLATE_TITLE,
     TEMPLATE_SUCCESS_BODY_TITLE,
     TEMPLATE_SUCCESS_TEXT,
