+++ project/events/constants/notification_types.py
