+++ project/events/services.py
@@ -18,13 +18,13 @@ from django.db import transaction
 from django.db.models import Q
 from django.db.models.query import QuerySet
 from django.utils import timezone
-from events.constant.notification_types import (
+from events.constants.notification_types import (
     NEW_USER_ON_THE_EVENT_NOTIFICATION_TYPE,
     RESPONSE_TO_THE_INVITE_TO_EVENT_NOTIFICATION_TYPE,
     RESPONSE_TO_THE_REQUEST_FOR_PARTICIPATION_NOTIFICATION_TYPE,
     USER_REMOVE_FROM_EVENT_NOTIFICATION_TYPE,
 )
-from events.constant.response_error import (
+from events.constants.response_error import (
     ALREADY_IN_EVENT_LIKE_SPECTATOR_ERROR,
     ALREADY_IN_EVENT_MEMBERS_LIST_ERROR,
     ALREADY_SENT_REQUEST_TO_PARTICIPATE_ERROR,
