+++ project/notifications/constants/notification_types.py
