+++ project/notifications/tasks.py
@@ -6,7 +6,7 @@ from authentication.models import User
 from channels.layers import get_channel_layer
 from config.celery import app
 from django.db.models import QuerySet
-from notifications.constant.notification_types import (
+from notifications.constants.notification_types import (
     CHANGE_MAINTENANCE_NOTIFICATION_TYPE,
 )
 from notifications.models import Notification
